requirements.txt
requests==2.31.0
python-dotenv==1.0.0
newspaper3k==0.2.8
nltk==3.8.1